vault-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vault-cli"
version = "0.1.0"
description = "CLI tool for managing a structured knowledge base alongside an Obsidian vault"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pymupdf4llm>=0.0.5",
    "trafilatura>=1.6.0",
    "python-frontmatter>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
vault = "vault.cli:app"

[tool.setuptools.packages.find]
include = ["vault*"]
